test: add MultipleChoiceCardStudyStrategyTests
Added the `MultipleChoiceCardStudyStrategyTests` class to test the `IsCorrectMultipleChoiceCardAnswer` method with various scenarios using `[Theory]` and `[InlineData]`. This change consolidates testing responsibilities by removing the previous implementation of the same test in `StudySessionsServiceTests.cs`, promoting better organization and clarity in the test suite.
